QuickFix of doubles

+++ Repository/Context/OverviewModel.cs
@@ -6,10 +6,10 @@
         public double tempAvg { get; set; } 
         public double tempMax { get; set; } 
         public int humiMin { get; set; } 
-        public int humiAvg { get; set; } 
+        public double humiAvg { get; set; } 
         public int humiMax { get; set; } 
         public int co2Min { get; set; } 
-        public int co2Avg { get; set; } 
+        public double co2Avg { get; set; } 
         public int co2Max { get; set; }
     }
 }
